tests/psr: Check for PSR entry more often and only for a second.
1 second interval is too long, check every 100ms.

PSR enabling shouldn't take 5 seconds on VLV/CHV and we should know if does
take more than a second. PSR should be enabled immediately after the pipe is
enabled on BDW+, so limit the checks to a 1 second total wait.

Signed-off-by: Dhinakaran Pandiyan <dhinakaran.pandiyan@intel.com>

squash with Check for PSR entry more often and only for a second.